Moves `Card` from `organisms` folder to `molecules`
