+++ packages/ui/src/molecules/Card/stories/Card.mdx
@@ -8,7 +8,7 @@ import CardActions from '../CardActions'
 # Card
 
 <Canvas>
-  <Story id="organisms-card--card" />
+  <Story id="molecules-card--card" />
 </Canvas>
 
 ## Components
