+++ themes/theme-b2c-tailwind/src/molecules/index.css
@@ -4,6 +4,7 @@
 @import './banner.css';
 @import './bullets.css';
 @import './breadcrumb.css';
+@import './card.css';
 @import './carousel.css';
 @import './dropdown.css';
 @import './form.css';
