--- themes/theme-b2c-tailwind/src/organisms/index.css
@@ -1,1 +1,0 @@
-@import './card.css';
